refactor: moved predictedHookActinoDate into notification/index
