+++ common/notification/hook.ts
@@ -5,9 +5,6 @@
 
 import { student as Student, pupil as Pupil } from '@prisma/client';
 import { getPupil, getStudent, User } from '../user';
-import { getNotifications } from './notification';
-import { prisma } from '../prisma';
-import { ActionID, ConcreteNotificationState } from './types';
 
 type NotificationHook = { fn: (user: User) => Promise<void>; description: string };
 
@@ -39,21 +36,3 @@ export const registerStudentHook = (hookID: string, description: string, hook: (
 
 export const registerPupilHook = (hookID: string, description: string, hook: (pupil: Pupil) => Promise<void>) =>
     registerHook(hookID, description, (user) => getPupil(user).then(hook));
-
-// Predicts when a hook will run for a certain user as caused by a certain action
-// i.e. 'When will a user by deactivated (hook) due to Certificate of Conduct reminders (action) ?'
-// Returns null if no date is known or hook was already triggered
-export async function predictedHookActionDate(action: ActionID, hookID: string, user: User): Promise<Date | null> {
-    const viableNotifications = ((await getNotifications()).get(action)?.toSend ?? []).filter((it) => it.hookID === hookID);
-
-    const possibleTrigger = await prisma.concrete_notification.findFirst({
-        where: {
-            state: ConcreteNotificationState.DELAYED,
-            userId: user.userID,
-            notificationID: { in: viableNotifications.map((it) => it.id) },
-        },
-        select: { sentAt: true },
-    });
-
-    return possibleTrigger?.sentAt;
-}
